Improve mobile view
This gets iOS view almost to where I want it, and the remaining issues
seem to be issues with Safari, not with my code:
- The input caret is mispositioned in some cases when typing, this
is a well-documented bug, see http://bit.ly/2AFTmtF I decided not to use
any of the workarounds there because changing off of fixed positioning
gave what I consider to be a worse experience
- If the user rotates to landscape and then back to portrait, the text
stays scaled to landscape until a refresh or new text comes in. I'd
honestly rather it not work in landscape at all since it's pretty
useless in that orientation, but that's not possible, so we do what we
can with what we have.

This addresses #15 but I need to check on some Android devices before
I close it.